assetsys: Add assetsys_mount_data() to mount from an archive data buffer
This change attempts to tackle #53 and adds a `assetsys_mount_data()` function to mount data from an archive data buffer. This paves the path to allow disabling STDIO, and cases in which the target already has the file loaded into memory.

It splits the `assetsys_mount()` into two different internal functions to save on code duplication. These functions may or may not be named correctly:
- `assetsys_internal_create_mount()`: Prepares a new mount for the assetsys instance
- `assetsys_internal_mount_zip(): Iterates through the mount system and indexes the files